requirements.txt
playwright
requests
beautifulsoup4
aiohttp
lxml
rich
mysql-connector-python
PyMySQL
Flask
gunicorn
